Move old components.
