+++ packages/notes/IndexSection.js
@@ -1,7 +1,7 @@
 'use strict';
 
 var Component = require('substance/ui/Component');
-var Dashboard = require('./Dashboard');
+var Dashboard = require('../dashboard/Dashboard');
 var Welcome = require('./Welcome');
 var EnterName = require('./EnterName');
