+++ packages/notes/RequestLogin.js
@@ -3,7 +3,7 @@
 var Component = require('substance/ui/Component');
 var Button = require('substance/ui/Button');
 var Input = require('substance/ui/Input');
-var Notification = require('./Notification');
+var Notification = require('../notification/Notification');
 
 function RequestLogin() {
   Component.apply(this, arguments);
@@ -16,8 +16,8 @@ RequestLogin.Prototype = function() {
 
     if (this.state.requested) {
       el.append(
-        $$('h1').append(this.i18n.t('sc-welcome.submitted-title')),
-        $$('p').append(this.i18n.t('sc-welcome.submitted-instructions'))
+        $$('h1').append(this.getLabel('sc-welcome.submitted-title')),
+        $$('p').append(this.getLabel('sc-welcome.submitted-instructions'))
       );
     } else {
       el.append(
@@ -34,7 +34,7 @@ RequestLogin.Prototype = function() {
       el.append(
         $$(Button, {
           disabled: !!this.state.loading // disable button when in loading state
-        }).append(this.i18n.t('sc-welcome.submit'))
+        }).append(this.getLabel('sc-welcome.submit'))
           .on('click', this._requestLoginLink)
       );
